+++ src/main/java/com/CuteNekoDragon/Core/common/data/SVOItemPrices.java
@@ -1,5 +1,6 @@
 package com.CuteNekoDragon.Core.common.data;
 
+import com.CuteNekoDragon.Core.common.data.items.SVOItems;
 import com.CuteNekoDragon.Core.utils.PriceUtil;
 
 public class SVOItemPrices {